+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## v1.2.0 - 2023-09-08
+### Added
+- `canonical_form(property: HplProperty)` function to `hpl.rewrite` module.
+
+### Changed
+- `HplAstObject.but()` returns the same object if no attributes would change.
+
 ## v1.1.2 - 2023-09-04
 ### Fixed
 - Sanity error when creating some types of predicates that was too strict.